requirements.txt
requests
python-dotenv
ass
